etl/pyproject.toml
[project]
name = "mise"
version = "0.1.0"
description = "Deterministic, Python-first ETL runner for the food graph"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "rich>=13.7.0",
  "PyYAML>=6.0",
  "jsonschema>=4.22",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
mise = "mise.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
